summary method for project configuration object

summary method for project configuration object

Usage

# S3 method for class 'bg_project_cfg'
summary(object, ...)

Arguments

object

The project configuration object (bg_project_cfg) to summarize.

...

additional parameters to summary (not used)

Value

Invisibly returns x after printing its contents. This function is called for its side effect of printing a formatted summary of the project configuration.